Improve call to nav-to-context in configure-desktop
The quest mainGoblin.configure-desktop can now return a contextId
to specify the context to open when the desktop is opened.
It allows to specify the initial context at run time instead of
configure it in the app configuration.

nav-to-context is not called if no contextId is defined
by configure-desktop or in the app config.