Fix behaviour when the locale is changed by the system or the user
The previous implementation was wrong in multiple ways. Here the
new and fixed behaviour:

The OS has always a locale. The first time (and if the application
doen't provide a way to change the locale) it's the OS locale which is
used for the app. If the user changes the OS locale, the app must change
the locale too on the next start.

Here the cases where the locale can be changed by the app (just for the
run or for the next runs too).

- The user wants an other locale that the OS, and only for one run. In
  this case the user can provide a command line argument `--locale=`
  in order to force a locale. The app starts with this locale. When the
  app is closed, the next start (without this argument) uses the OS
  locale.
- The user wants an other locale everytime (different of the OS). Here
  it's a bit different, because the user select the new locale from the
  app, then this locale is saved in the clientSession file. When the app
  is closed and open again, this (saved) locale is used instead of the
  OS.

For the second point there is something special. It's fine while the
user chooses a locale different of the OS. But when it's the same, it's
just wrong to save the locale in the clientSession file. When the user
chooses the same locale that the OS, then the entry in the clientSession
file is remove and the OS fallback is used instead.